#68 introduces critical script identification mechanism
To mark a script as system critical add the prefix *CRIT* after the script number.

To mark a script as non-critical it is recommended to add the prefix *NORM* even though it has no programmatic effect, it results in a better overview of all scripts.

Note: Errors in critical scripts will result in a jenkins shutdown.