#94 Set custom Certificate store for `curl`
Additional certificates may be read during from the etcd, but adding it to the
original certificate store is out of reach because the container runs as
Jenkins user. In contrast, actions to the original certificate store must be
run as root.

`curl` uses by default the original certificate store
`/etc/ssl/certs/ca-certificates.crt`. When natively called, curl will fail for
https sites that use self-signed certificates which are not imported to the
original certificate store. Jenkins custom certificate store resides outside
`/etc/` because the Jenkins dogu runs as non-root for security reasons.

The solution is to add a curl rc file with Jenkins's custom certificate store
so that even native `curl` calls will find additional certificates.