#94 Add single additional certificates for subversion
Subversion is able to add additional certificates so it can communicate with
external services that secure their network stream with self-signed certificates.
In order to do so, Subversion needs to register each certificate inside a single
file as well as a Subversion configuration which refers to said certificate
files. The certificate format must be PEM (which is already the used format from
the `etcd`).

`splitCertificates()` splits the multiple certificates from single etcd keys
with help from `csplit`. The result are one or more certificate files which
automatically receive a zero-based counter suffix, f. i.:
cert-my.fqdn.com-00, ...-01, ...-02 and so on.

The concrete file names cannot be fetched from stdout because `csplit` creates
the files right away. So instead of trying to read the actually created files
from the file system, `addCertificatePathToSubversionSSLConfig()` creates the
zero-based file references retrospectively by help of the number of certificates
per key. This is because f. i. two certificates should always lead to two
certificate files.