Fix a bug with older versions of the matrix-auth plugin.
The init050CRITsetauthorization script from v2.332.1-2 only works if the version of the matrix-auth plugin is > 3.0.
The bug that was fixed in the last version does only affect jenkins instances where the matrix-auth plugin is > v3.0. Versions prior to that don't have that bug and therefore the bugfix crashes the jenkins during startup.