Fix a bug with older versions of the matrix-auth plugin.
Resloves #113.
The init050CRITsetauthorization script from v2.332.1-2 only worked if the version of the matrix-auth plugin is > 3.0.
Therefore the installPluigns script now updates the matrix-auth plugin if it is below version 3.0. This is done because the plugin versions after v3.0 refactored permission-groups handling.

This has already been solved similarly for the CAS plugin in an earlier development. The way of the CAS plugin was adapted to the new implementation.