Fixed startup loop when matrix-auth plugin is installed below v3.0
The installPluigns script now updates the matrix-auth plugin if it is below version 3.0. This is done because the plugin versions after v3.0 refactored permission-groups handling.

This has already been solved similarly for the CAS plugin in an earlier development. The way of the CAS plugin was adapted to the new implementation.