#166 Remove Third-Party glibc compat layer in favor of gcompat
Apparently, Jenkins depends on glibc but Alpine comes by default with musl.
In order to solve this matter, a third party compatibility layer `sgerrand` was
downloaded and installed. With Alpine 3.20 this seems to err with this message:

ERROR: gcompat-1.1.0-r4: trying to overwrite lib/ld-linux-x86-64.so.2 owned by glibc-2.35-r1.

Since Alpine has its [go-to compatibility layer] `gcompat`, this commit removes
the need for a third party software.

[go-to compatibility layer]: https://wiki.alpinelinux.org/wiki/Running_glibc_programs

Quote:
```
gcompat

gcompat is the go-to compatibility layer for Alpine users.

apk add gcompat

After that you run your binaries as normal.
```